PARQUET-324: row count incorrect if data file has more than 2^31 rows
Need to change numRows counter from int to long to account for input files with more than 2^31 rows.

Author: Thomas Friedrich <tfriedr@us.ibm.com>

Closes #233 from tfriedr/parquet-324 and squashes the following commits:

0120205 [Thomas Friedrich] change numRows from int to long

+++ parquet-hadoop/src/main/java/org/apache/parquet/format/converter/ParquetMetadataConverter.java
@@ -89,7 +89,7 @@ public class ParquetMetadataConverter {
   public FileMetaData toParquetMetadata(int currentVersion, ParquetMetadata parquetMetadata) {
     List<BlockMetaData> blocks = parquetMetadata.getBlocks();
     List<RowGroup> rowGroups = new ArrayList<RowGroup>();
-    int numRows = 0;
+    long numRows = 0;
     for (BlockMetaData block : blocks) {
       numRows += block.getRowCount();
       addRowGroup(parquetMetadata, rowGroups, block);
